Parameters functions, weather types added.
Temperature, SoilPh, Nutrient, Humidity are functions which are wheather
specific. When wheather  changes, functions also change. They also
depend on SystemTimer elapsed time - you can set one cycle(2 min. for
ex.) - during the cycle functions vary in different way.

+++ Plants/MessagesCreators/MeasuringMessageCreator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Planting.MeasuringsProviding;
-using Planting.PlantRequirements;
+using Planting.Messenging;
+using Planting.PlantsRequirements;
 
 namespace Planting.MessagesCreators
 {
@@ -8,9 +9,9 @@ namespace Planting.MessagesCreators
     {
         public MeasurableParameter MeasurableParameter { get; set; }
         public string PlantsAreaId { get; private set; }
-        public int ParameterValue { get; private set; }
+        public double ParameterValue { get; private set; }
 
-        public MeasuringMessageCreator(MeasurableParameter measurableValue, string plantsAreaId, int measuringValue)
+        public MeasuringMessageCreator(MeasurableParameter measurableValue, string plantsAreaId, double measuringValue)
         {
             MeasurableParameter = measurableValue;
             PlantsAreaId = plantsAreaId;
